feat: move bundle and add table to print receipt
