+++ inn/inn_hotels/page/pos_extended/pos_extended.js
@@ -19,10 +19,50 @@ frappe.pages['pos-extended'].on_page_load = function (wrapper) {
 
 	document.getElementsByTagName("head")[0].appendChild(link);
 
-	frappe.require("point-of-sale.bundle.js", function () {
+	frappe.require(["point-of-sale.bundle.js", "inn-pos.bundle.js"], function () {
 		inn.PointOfSale.Controller = class MyPosController extends erpnext.PointOfSale.Controller {
 			constructor(wrapper) {
 				super(wrapper);
+			}
+			init_payments() {
+				this.payment = new erpnext.PointOfSale.Payment({
+					wrapper: this.$components_wrapper,
+					events: {
+						get_frm: () => this.frm || {},
+
+						get_customer_details: () => this.customer_details || {},
+
+						toggle_other_sections: (show) => {
+							if (show) {
+								this.item_details.$component.is(':visible') ? this.item_details.$component.css('display', 'none') : '';
+								this.item_selector.toggle_component(false);
+							} else {
+								this.item_selector.toggle_component(true);
+							}
+						},
+
+						submit_invoice: () => {
+							this.frm.savesubmit()
+								.then((r) => {
+									this.toggle_components(false);
+									this.order_summary.toggle_component(true);
+									this.order_summary.load_summary_of(this.frm.doc, true);
+
+									frappe.call({
+										method: "inn.inn_hotels.page.pos_extended.pos_extended.clean_table_number",
+										args: {
+											invoice_name: this.frm.doc.name
+										}
+									})
+
+									frappe.show_alert({
+										indicator: 'green',
+										message: __('POS invoice {0} created succesfully', [r.doc.name])
+									});
+								});
+						}
+					}
+				});
 			}
 
 			save_draft_invoice() {
@@ -71,7 +111,6 @@ frappe.pages['pos-extended'].on_page_load = function (wrapper) {
 			}
 
 			prepare_menu() {
-				this.page.clear_menu();
 				this.page.add_menu_item(("Open Form View"), this.open_form_view.bind(this), false, 'Ctrl+F');
 				this.page.add_menu_item(("Toggle Recent Orders"), this.toggle_recent_order.bind(this), false, 'Ctrl+O');
 				this.page.add_menu_item(("Save as Draft"), this.save_draft_invoice.bind(this), false, 'Ctrl+S');
@@ -100,166 +139,77 @@ frappe.pages['pos-extended'].on_page_load = function (wrapper) {
 							this.customer_details = details;
 							// will add/remove LP payment method
 							this.payment.render_loyalty_points_payment_mode();
+						},
+
+						print_captain_order: () => this.print_captain_order(),
+						print_table_order: () => this.print_table_order(),
+						print_bill: () => this.print_bill(),
+					}
+				})
+			}
+
+			print_table_order() {
+				console.log("table order")
+			}
+
+			print_captain_order() {
+				console.log("captain order")
+			}
+
+			print_bill() {
+
+
+				// let w = window.open(frappe.urllib.get_full_url("/printview?"
+				// 	+ "doctype=" + encodeURIComponent("Inn Restaurant Ongoing Order")
+				// 	+ "&name=" + encodeURIComponent(frm.doc.name)
+				// 	+ "&format=" + encodeURIComponent("Captain Order")
+				// 	+ "&trigger_print=1"
+				// 	+ "&no_letterhead=0"
+				// ));
+			}
+
+			init_order_summary() {
+				this.order_summary = new inn.PointOfSale.PosExtendPastOrderSummary({
+					wrapper: this.$components_wrapper,
+					events: {
+						get_frm: () => this.frm,
+
+						process_return: (name) => {
+							this.recent_order_list.toggle_component(false);
+							frappe.db.get_doc('POS Invoice', name).then((doc) => {
+								frappe.run_serially([
+									() => this.make_return_invoice(doc),
+									() => this.cart.load_invoice(),
+									() => this.item_selector.toggle_component(true)
+								]);
+							});
+						},
+						edit_order: (name) => {
+							this.recent_order_list.toggle_component(false);
+							frappe.run_serially([
+								() => this.frm.refresh(name),
+								() => this.frm.call('reset_mode_of_payments'),
+								() => this.cart.load_invoice(),
+								() => this.item_selector.toggle_component(true)
+							]);
+						},
+						delete_order: (name) => {
+							frappe.model.delete_doc(this.frm.doc.doctype, name, () => {
+								this.recent_order_list.refresh_list();
+							});
+						},
+						new_order: () => {
+							frappe.run_serially([
+								() => frappe.dom.freeze(),
+								() => this.make_new_invoice(),
+								() => this.item_selector.toggle_component(true),
+								() => frappe.dom.unfreeze(),
+							]);
 						}
 					}
 				})
 			}
 		};
-
-
-		inn.PointOfSale.PosExtendItemCart = class PosExtendItemCart extends erpnext.PointOfSale.ItemCart {
-			constructor({ wrapper, events, settings }) {
-				super({ wrapper, events, settings });
-			}
-			init_child_components() {
-				this.init_customer_selector();
-				this.init_table_selector();
-				this.init_cart_components();
-			}
-
-			init_table_selector() {
-				this.$component.append(
-					`<div class="table-section"></div>`
-				)
-				this.$table_section = this.$component.find('.table-section');
-				this.make_table_selector();
-			}
-
-
-			make_table_selector() {
-				this.$table_section.html(`<div class="table-field"></div>`)
-				const me = this;
-				this.table_field = frappe.ui.form.make_control({
-					df: {
-						label: __('No Table'),
-						fieldtype: 'Link',
-						options: 'Inn Point Of Sale Table',
-						placeholder: __('Search table number'),
-						get_query: function () {
-							return {
-								filters: [
-									["Inn Point Of Sale Table", "status", "=", "Empty"]
-								]
-							}
-						},
-						onchange: function () {
-							me["table_number"] = this.get_value()
-						}
-					},
-					parent: this.$table_section.find('.table-field'),
-					render_input: true,
-				});
-				this.table_field.toggle_label(false);
-			}
-
-			make_cart_totals_section() {
-				this.$totals_section = this.$component.find('.cart-totals-section');
-
-				this.$totals_section.append(
-					`<div class="add-discount-wrapper">
-						${this.get_discount_icon()} ${__('Add Discount')}
-					</div>
-					<div class="item-qty-total-container">
-						<div class="item-qty-total-label">${__('Total Items')}</div>
-						<div class="item-qty-total-value">0.00</div>
-					</div>
-					<div class="net-total-container">
-						<div class="net-total-label">${__("Net Total")}</div>
-						<div class="net-total-value">0.00</div>
-					</div>
-					<div class="taxes-container"></div>
-					<div class="grand-total-container">
-						<div>${__('Grand Total')}</div>
-						<div>0.00</div>
-					</div>
-					<div class="print-order-section">
-						<div class="caption-order-btn" data-button-value="captain-order">${__('Captain Order')}</div>
-						<div class="table-order-btn" data-button-value="table-order">${__('Table Order')}</div>
-					</div>
-					<div class="checkout-btn">${__('Checkout')}</div>
-					<div class="edit-cart-btn">${__('Edit Cart')}</div>`
-				)
-
-				this.$add_discount_elem = this.$component.find(".add-discount-wrapper");
-			}
-
-			highlight_checkout_btn(toggle) {
-				if (toggle) {
-					this.$add_discount_elem.css('display', 'flex');
-					this.$cart_container.find('.checkout-btn').css({
-						'background-color': 'var(--blue-500)'
-					});
-					this.$cart_container.find('.caption-order-btn').css({
-						'background-color': 'var(--gray-800)'
-					});
-					this.$cart_container.find('.table-order-btn').css({
-						'background-color': 'var(--gray-800)'
-					});
-				} else {
-					this.$add_discount_elem.css('display', 'none');
-					this.$cart_container.find('.checkout-btn').css({
-						'background-color': 'var(--blue-200)'
-					});
-					this.$cart_container.find('.caption-order-btn').css({
-						'background-color': 'var(--gray-300)'
-					});
-					this.$cart_container.find('.table-order-btn').css({
-						'background-color': 'var(--gray-300)'
-					});
-				}
-			}
-		}
-
-		erpnext.PointOfSale.PastOrderSummary = class PosExtendPastOrderSummary extends erpnext.PointOfSale.PastOrderSummary {
-			constructor({ wrapper, events }) {
-				super({ wrapper, events })
-			}
-
-			async attach_document_info(doc) {
-				await frappe.db.get_value('Customer', this.doc.customer, 'email_id').then(({ message }) => {
-					doc.customer_email = message.email_id || '';
-				});
-
-				const upper_section_dom = this.get_upper_section_html(doc);
-				this.$upper_section.html(upper_section_dom);
-			}
-
-			get_upper_section_html(doc) {
-				const { status } = doc;
-				let indicator_color = '';
-
-				in_list(['Paid', 'Consolidated'], status) && (indicator_color = 'green');
-				status === 'Draft' && (indicator_color = 'red');
-				status === 'Return' && (indicator_color = 'grey');
-
-				var table_number = undefined
-
-				frappe.call({
-					method: "inn.inn_hotels.page.pos_extended.pos_extended.get_table_number",
-					args: {
-						invoice_name: doc.name
-					},
-					async: false,
-					callback: function (r) {
-						table_number = r.message
-					}
-				})
-
-				return `<div class="left-section">
-							<div class="customer-name">${doc.customer}</div>
-							<div class="customer-email">${doc.customer_email}</div>
-							<div class="table-number">${__('Table')}: ${table_number}</div>
-							<div class="cashier">${__('Sold by')}: ${doc.owner}</div>
-						</div>
-						<div class="right-section">
-							<div class="paid-amount">${format_currency(doc.paid_amount, doc.currency)}</div>
-							<div class="invoice-name">${doc.name}</div>
-							<span class="indicator-pill whitespace-nowrap ${indicator_color}"><span>${doc.status}</span></span>
-						</div>`;
-			}
-		}
-
-
 
 		wrapper.pos = new inn.PointOfSale.Controller(wrapper);
 		window.cur_pos = wrapper.pos;
